requirements.txt
pypdf==6.6.0
gradio==6.3.0
chromadb==1.4.1
sentence-transformers==5.2.0  
langchain==1.2.4
langchain-groq==1.1.1
langchain-community==0.4.1
langchain-huggingface==1.2.0
python-dotenv==1.2.1
reportlab==4.4.10